uluwatu 2nd image

+++ uluwatu/build-tar.sh
@@ -1,10 +1,15 @@
+rm -rf uluwatu.tar uluwatu
 gpg --recv-key 91824AB09120A048
-ln -sf /usr/share/debootstrap/scripts/sid \
+if [ -f "/usr/share/debootstrap/scripts/uluwatu" ]
+then
+ ln -sf /usr/share/debootstrap/scripts/sid \
    /usr/share/debootstrap/scripts/uluwatu
-debootstrap --keyring=~/.gnupg/pubring.gpg --variant=minbase \
+fi
+debootstrap --keyring=/root/.gnupg/pubring.kbx --variant=minbase \
    --components=main --include=inetutils-ping,iproute2 \
    uluwatu uluwatu http://arsip-dev.blankonlinux.or.id/blankon
 cd uluwatu
 tar cf ../uluwatu.tar *
 cd ..
 xz -9 uluwatu.tar
+rm -rf uluwatu
